goy_app/requirements.txt
asgiref==3.8.1
confluent-kafka==2.5.0
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
